Bot/chat_bot/requirements.txt
aiohttp
cryptography==45.0.3
PyJWT==2.10.1
botbuilder-core==4.16.2
botbuilder-dialogs==4.16.2
botbuilder-schema==4.16.2

